Move controller functionality to fint-core-consumer-shared
