+++ src/main/java/no/fintlabs/consumer/admin/AdminController.java
@@ -3,8 +3,8 @@ package no.fintlabs.consumer.admin;
 import no.fintlabs.cache.CacheManager;
 import no.fintlabs.consumer.config.RestEndpoints;
 import no.fintlabs.core.consumer.shared.ConsumerProps;
-import no.fintlabs.core.consumer.shared.ConsumerService;
-import no.fintlabs.core.consumer.shared.admin.BaseAdminController;
+import no.fintlabs.core.consumer.shared.admin.ConsumerAdminController;
+import no.fintlabs.core.consumer.shared.resource.ConsumerService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.MediaType;
 import org.springframework.web.bind.annotation.CrossOrigin;
@@ -16,7 +16,7 @@ import java.util.Collection;
 @RestController
 @CrossOrigin
 @RequestMapping(value = RestEndpoints.ADMIN, produces = MediaType.APPLICATION_JSON_VALUE)
-public class AdminController extends BaseAdminController {
+public class AdminController extends ConsumerAdminController {
 
     @Autowired(required = false)
     private Collection<ConsumerService<?>> consumerServices;
