--- src/main/java/no/fintlabs/consumer/exception/EntityNotFoundException.java
@@ -1,7 +1,0 @@
-package no.fintlabs.consumer.exception;
-
-public class EntityNotFoundException extends RuntimeException {
-    public EntityNotFoundException(String message) {
-        super(message);
-    }
-}
