+++ src/main/java/no/fintlabs/consumer/fravar/FravarService.java
@@ -8,7 +8,7 @@ import no.fintlabs.cache.Cache;
 import no.fintlabs.cache.CacheManager;
 import no.fintlabs.cache.packing.PackingTypes;
 import no.fintlabs.core.consumer.shared.ConsumerProps;
-import no.fintlabs.core.consumer.shared.ConsumerService;
+import no.fintlabs.core.consumer.shared.resource.ConsumerService;
 import org.apache.kafka.clients.consumer.ConsumerRecord;
 import org.springframework.stereotype.Service;
 
@@ -48,7 +48,8 @@ public class FravarService extends ConsumerService<FravarResource> {
         //log.info("The cache now containes " + this.getCacheSize() + " elements.");
     }
 
-    public Optional<FravarResource> getFravarBySystemId(String systemId) {
+    @Override
+    public Optional<FravarResource> getBySystemId(String systemId) {
         return getCache().getLastUpdatedByFilter(systemId.hashCode(),
                 (resource) -> Optional
                         .ofNullable(resource)
